Switch to use the new precompiled interface
